only save/use iops for io1 volumes

+++ pkg/backup/volume_snapshot_action_test.go
@@ -31,7 +31,7 @@ import (
 )
 
 func TestVolumeSnapshotAction(t *testing.T) {
-	iops := 1000
+	iops := int64(1000)
 
 	tests := []struct {
 		name                  string
